Add media service upload operation.
