--- components/org.wso2.carbon.identity.image/src/main/java/org/wso2/carbon/identity/image/file/FileBasedStorageSystemFactory.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wso2.carbon.identity.image.file;
-
-import org.wso2.carbon.identity.image.StorageSystem;
-import org.wso2.carbon.identity.image.StorageSystemFactory;
-
-/**
- * Factory class related to producing FileBasedStorageSystemImpl objects.
- */
-public class FileBasedStorageSystemFactory extends StorageSystemFactory {
-
-    public static final String FILE = "org.wso2.carbon.identity.image.file.FileBasedStorageSystemImpl";
-
-    @Override
-    public StorageSystem getInstance() {
-
-        StorageSystem fileBasedStorageSystemImpl = new FileBasedStorageSystemImpl();
-        return fileBasedStorageSystemImpl;
-    }
-
-    @Override
-    public String getStorageType() {
-
-        return FILE;
-    }
-
-}
