--- components/org.wso2.carbon.identity.image/src/main/java/org/wso2/carbon/identity/image/util/StorageSystemConstants.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.wso2.carbon.identity.image.util;
-
-/**
- * Class to hold constants.
- */
-public class StorageSystemConstants {
-
-    public static final String ID_SEPERATOR = "_";
-    public static final String SYSTEM_PROPERTY_CARBON_HOME = "carbon.home";
-    public static final String IMAGE_STORE = "repository/images";
-    public static final String IDP = "idp";
-    public static final String SP = "app";
-    public static final String USER = "user";
-    public static final String DEFAULT = "default";
-    public static final String CONFIGURABLE_UPLOAD_LOCATION = "upload.location";
-    public static final String FILE_CREATION_TIME_ATTRIBUTE = "creationTime";
-}
