+++ components/org.wso2.carbon.identity.media/src/main/java/org/wso2/carbon/identity/media/file/FileBasedStorageSystemImpl.java
