+++ components/org.wso2.carbon.identity.media/src/main/java/org/wso2/carbon/identity/media/util/StorageSystemConstants.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.wso2.carbon.identity.media.util;
+
+/**
+ * Class to hold constants.
+ */
+public class StorageSystemConstants {
+
+    public static final String ID_SEPERATOR = "_";
+    public static final String SYSTEM_PROPERTY_CARBON_HOME = "carbon.home";
+    public static final String MEDIA_STORE = "repository/media/";
+    public static final String IDP = "idp";
+    public static final String SP = "app";
+    public static final String USER = "user";
+    public static final String DEFAULT = "default";
+    public static final String CONFIGURABLE_UPLOAD_LOCATION = "upload.location";
+    public static final String FILE_CREATION_TIME_ATTRIBUTE = "creationTime";
+    public static final String METADATA_FILE_SUFFIX = "_meta";
+    public static final String METADATA_FILE_EXTENSION = ".json";
+    public static final String METADATA_FILE_NAME = "fileName";
+    public static final String METADATA_FILE_CONTENT_TYPE = "fileContentType";
+    public static final String METADATA_FILE_TAG = "fileTag";
+    public static final String METADATA_FILE_CREATED_TIME = "fileCreatedTime";
+    public static final String METADATA_FILE_LAST_ACCESSED_TIME = "fileLastAccessedTime";
+    public static final String METADATA_FILE_SECURITY = "fileSecurity";
+    public static final String FILE_SECURITY_ALLOWED_ALL = "allowedAll";
+    public static final String FILE_SECURITY_ALLOWED_USERS = "allowedUsers";
+    public static final String FILE_SECURITY_ALLOWED_SCOPES = "allowedScopes";
+
+}
