+++ components/org.wso2.carbon.identity.media/src/test/resources/profilepic.png
