+++ features/org.wso2.carbon.identity.media.server.feature/pom.xml
@@ -24,18 +24,17 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.identity.content.repository.feature</artifactId>
+    <artifactId>org.wso2.carbon.identity.media.server.feature</artifactId>
     <packaging>pom</packaging>
-    <name>WSO2 Carbon - Identity Content Repository Feature</name>
+    <name>WSO2 Carbon - Identity Media Service Feature</name>
     <url>http://wso2.org</url>
-    <description>This feature contains the core bundles for identity content repository.</description>
+    <description>This feature contains the core bundles for media service.</description>
 
     <dependencies>
         <dependency>
             <groupId>org.wso2.carbon.identity.content.repository</groupId>
-            <artifactId>org.wso2.carbon.identity.media.server.feature</artifactId>
+            <artifactId>org.wso2.carbon.identity.media</artifactId>
             <version>${project.version}</version>
-            <type>zip</type>
         </dependency>
     </dependencies>
 
@@ -53,22 +52,22 @@
                             <goal>p2-feature-gen</goal>
                         </goals>
                         <configuration>
-                            <id>org.wso2.carbon.identity.content.repository</id>
+                            <id>org.wso2.carbon.identity.media.server</id>
                             <propertiesFile>../etc/feature.properties</propertiesFile>
-                            <importFeatures>
-                                <importFeatureDef>
-                                    org.wso2.carbon.core:compatible:${carbon.kernel.feature.version}
-                                </importFeatureDef>
-                            </importFeatures>
-                            <includedFeatures>
-                                <incincludedFeatureDef>
-                                    org.wso2.carbon.identity.content.repository:org.wso2.carbon.identity.media.server.feature
-                                </incincludedFeatureDef>
-                            </includedFeatures>
+                            <adviceFile>
+                                <properties>
+                                    <propertyDef>org.wso2.carbon.p2.category.type:server</propertyDef>
+                                </properties>
+                            </adviceFile>
+                            <bundles>
+                                <bundleDef>org.wso2.carbon.identity.content.repository:org.wso2.carbon.identity.media
+                                </bundleDef>
+                            </bundles>
                         </configuration>
                     </execution>
                 </executions>
             </plugin>
         </plugins>
     </build>
+
 </project>
